Store serialized access token as string
This change modifies Doctrine's column schema for the serialized
access token so it's stored as a string, not as an "array".

This doesn't change any actual data in the database, but changes the
set- and getSerializedAccesstoken() methods which now expect a string
instead of an array.

This change also improves test coverage of the Authorization class.