Remove OAuth client secret from Authorization records
Remove the client secret from the Authorization table – for now at
least. We don't need to store the client secret, because if OAuth is
used with client credentials flow, the client secret is likely
available somewhere lese (for example via a setting or an application
specific storage) and when authorization code flow is used, the secret
is only needed when authorization is started and does not have to be
stored.

We may need the client secret for refreshing a token, but this is not
correctly implemented at this time, so we may need to solved this at
a later point in time.